+++ packages/terra-navigation-side-menu/CHANGELOG.md
@@ -4,3 +4,4 @@ ChangeLog
 Unreleased
 -----------------
 * Initial stable release
+* Update focus styling to match ux ask